BACKPORT: sched/fair: Add lag based placement
With the introduction of avg_vruntime, it is possible to approximate
lag (the entire purpose of introducing it in fact). Use this to do lag
based placement over sleep+wake.

Specifically, the FAIR_SLEEPERS thing places things too far to the
left and messes up the deadline aspect of EEVDF.

BUG=b:308209790
TEST=Manual tested on DUT

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Link: https://lore.kernel.org/r/20230531124603.794929315@infradead.org
Signed-off-by: Youssef Esmat <youssefesmat@chromium.org>

Change-Id: Ie21f7d37d0dce29701befb2afa2591746c29ccb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5089893
Reviewed-by: Vineeth Pillai <vineethrp@google.com>
Commit-Queue: Youssef Esmat <youssefesmat@chromium.org>
Tested-by: Vineeth Pillai <vineethrp@google.com>