tensor_aio: Use the highest devfreq frequency scaled by CPU capacity
Not all CPUs are created equal. They have different frequency tables and
capacities, so boosting devfreq to max when any CPU is at the maxmium
frequency could be excessive.

Assume that devfreq devices only need to be at their maximum frequency when
the biggest core is at the highest frequency, and scale accordingly from
there.

Signed-off-by: Kazuki H <kazukih0205@gmail.com>